+++ src/media/PortableSegmentRecorder.h
@@ -8,8 +8,13 @@
 #include <QTimer>
 
 #include <chrono>
+#include <memory>
 
 class QTemporaryFile;
+
+namespace LastFrame::Platform {
+class WindowsDesktopCapture;
+}
 
 namespace LastFrame::Media {
 
@@ -61,6 +66,7 @@ private:
     [[nodiscard]] int nextSegmentNumber() const;
     [[nodiscard]] QString escapeConcatPath(const QString& path) const;
     [[nodiscard]] QString selectedMonitorLabel() const;
+    [[nodiscard]] bool prepareNativeCapture();
     void discoverMicrophoneDevice();
     void startProcess(bool withAudio, bool announceStarted = true);
     void finishExport(ExportJob* job, int exitCode, QProcess::ExitStatus status);
@@ -78,10 +84,15 @@ private:
     QString microphoneDeviceName_;
     bool useDesktopDuplication_ = true;
     bool attemptedDesktopDuplicationFallback_ = false;
+    bool useNativeCapture_ = false;
+    bool attemptedNativeCaptureFallback_ = false;
     bool useSoftwareEncoder_ = false;
     bool attemptedEncoderFallback_ = false;
     LastFrame::Core::RateLimiter rateLimiter_{3, std::chrono::seconds(1), std::chrono::seconds(5)};
     QList<ExportJob*> exports_;
+#if defined(Q_OS_WIN)
+    std::unique_ptr<LastFrame::Platform::WindowsDesktopCapture> nativeCapture_;
+#endif
 };
 
 } // namespace LastFrame::Media